refactor: move test utils, apply eslint on cypress
