+++ .eslintrc
@@ -22,6 +22,8 @@
     // remove when possible
     "@typescript-eslint/no-explicit-any": "off",
     // disable the rule for all files
+    "no-restricted-syntax": "off",
+    // disable the rule for all files
     "@typescript-eslint/explicit-module-boundary-types": "off",
     "jsx-a11y/anchor-is-valid": [
       "error",
@@ -73,6 +75,13 @@
       "rules": {
         "@typescript-eslint/explicit-module-boundary-types": ["error"]
       }
+    },
+    {
+      // enable the rule specifically for src files
+      "files": ["src/**/*.js", "src/**/*.tsx", "src/**/*.ts"],
+      "rules": {
+        "no-restricted-syntax": ["error"]
+      }
     }
   ],
   "settings": {
@@ -89,5 +98,5 @@
       }
     }
   },
-  "ignorePatterns": ["cypress/*"]
+  // "ignorePatterns": ["cypress/*"]
 }
