+++ cypress/e2e/item/chatbox/chatbox.cy.js
@@ -1,4 +1,6 @@
+// eslint-disable-next-line import/no-extraneous-dependencies
 import { WebSocket } from '@graasp/websockets/test/mock-client';
+
 import { buildItemPath } from '../../../../src/config/paths';
 import {
   CHATBOX_ID,
