+++ cypress/e2e/item/edit/editFile.cy.js
@@ -1,10 +1,10 @@
 import { DEFAULT_ITEM_LAYOUT_MODE } from '../../../../src/config/constants';
+import { HOME_PATH, buildItemPath } from '../../../../src/config/paths';
 import { ITEM_LAYOUT_MODES } from '../../../../src/enums';
-import { buildItemPath, HOME_PATH } from '../../../../src/config/paths';
 import { IMAGE_ITEM_DEFAULT, VIDEO_ITEM_S3 } from '../../../fixtures/files';
 import { EDITED_FIELDS } from '../../../fixtures/items';
 import { EDIT_ITEM_PAUSE } from '../../../support/constants';
-import { editCaptionFromViewPage, editItem } from './utils';
+import { editCaptionFromViewPage, editItem } from '../../../support/editUtils';
 
 describe('Edit File', () => {
   beforeEach(() => {
