+++ cypress/e2e/item/edit/editFolder.cy.js
@@ -1,16 +1,16 @@
 import { DEFAULT_ITEM_LAYOUT_MODE } from '../../../../src/config/constants';
+import { HOME_PATH, buildItemPath } from '../../../../src/config/paths';
+import {
+  ITEM_FORM_CONFIRM_BUTTON_ID,
+  buildEditButtonId,
+} from '../../../../src/config/selectors';
 import { ITEM_LAYOUT_MODES } from '../../../../src/enums';
-import { buildItemPath, HOME_PATH } from '../../../../src/config/paths';
 import { EDITED_FIELDS, SAMPLE_ITEMS } from '../../../fixtures/items';
 import {
   EDIT_ITEM_PAUSE,
   TABLE_ITEM_RENDER_TIME,
 } from '../../../support/constants';
-import { editItem, editCaptionFromViewPage } from './utils';
-import {
-  buildEditButtonId,
-  ITEM_FORM_CONFIRM_BUTTON_ID,
-} from '../../../../src/config/selectors';
+import { editCaptionFromViewPage, editItem } from '../../../support/editUtils';
 
 describe('Edit Folder', () => {
   describe('List', () => {
