+++ cypress/e2e/item/edit/editLink.cy.js
@@ -1,13 +1,13 @@
 import { DEFAULT_ITEM_LAYOUT_MODE } from '../../../../src/config/constants';
+import { HOME_PATH, buildItemPath } from '../../../../src/config/paths';
 import { ITEM_LAYOUT_MODES } from '../../../../src/enums';
-import { buildItemPath, HOME_PATH } from '../../../../src/config/paths';
 import { EDITED_FIELDS } from '../../../fixtures/items';
 import { GRAASP_LINK_ITEM, YOUTUBE_LINK_ITEM } from '../../../fixtures/links';
 import {
   EDIT_ITEM_PAUSE,
   ITEM_LOADING_PAUSE,
 } from '../../../support/constants';
-import { editCaptionFromViewPage, editItem } from './utils';
+import { editCaptionFromViewPage, editItem } from '../../../support/editUtils';
 
 describe('Edit Link', () => {
   beforeEach(() => {
