+++ cypress/e2e/item/view/viewFile.cy.js
@@ -1,7 +1,7 @@
 import { DEFAULT_ITEM_LAYOUT_MODE } from '../../../../src/config/constants';
-import { ITEM_LAYOUT_MODES } from '../../../../src/enums';
 import { HOME_PATH } from '../../../../src/config/paths';
 import { buildItemsTableRowIdAttribute } from '../../../../src/config/selectors';
+import { ITEM_LAYOUT_MODES } from '../../../../src/enums';
 import {
   IMAGE_ITEM_DEFAULT,
   IMAGE_ITEM_S3,
@@ -10,7 +10,7 @@ import {
   VIDEO_ITEM_DEFAULT,
   VIDEO_ITEM_S3,
 } from '../../../fixtures/files';
-import { expectFileViewScreenLayout } from './utils';
+import { expectFileViewScreenLayout } from '../../../support/viewUtils';
 
 describe('Files', () => {
   describe('default server', () => {
